fix box generation

+++ src/bounding_volume_hierarchy.h
@@ -41,7 +41,7 @@ public:
     BoundingVolumeHierarchy(Scene* pScene);
 
     // construction helper, returns index of last added node
-    size_t constructorHelper(std::vector<TriangleOrNode>& triangles, int whichAxis);
+    size_t constructorHelper(std::vector<TriangleOrNode>& triangles, size_t whichAxis);
 
     // Return how many levels there are in the tree that you have constructed.
     [[nodiscard]] int numLevels() const;
